backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# LLM Integration
litellm==1.52.0

# MCP Integration
mcp==1.1.2

# Memory Management
mem0ai==0.1.30
sentence-transformers==3.3.1
qdrant-client==1.12.1
sentence-transformers==3.3.1
qdrant-client==1.12.1

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Security
cryptography==44.0.0
python-dotenv==1.0.1

# Data Processing
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.5

# HTTP Client
httpx==0.27.2

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
